Downloads/agrimind-ai/agrimind/backend/services/disease_service.py
"""
AgriMind AI - Disease Detection Service
Heuristic disease analysis with deterministic scoring
"""

import os
from PIL import Image
import numpy as np
from services.translation_service import translate_text


DISEASE_KNOWLEDGE = {
    "Healthy": {
        "description": "The plant appears healthy with no strong visible signs of disease or stress.",
        "treatment": "No treatment required. Continue regular monitoring and balanced crop care.",
        "prevention": "Maintain proper irrigation, balanced fertilization, and field hygiene.",
        "severity": "None",
        "affected_crops": ["All crops"]
    },
    "Bacterial Leaf Blight": {
        "description": "Leaf margins show water-soaked yellowing that later becomes brown and dry.",
        "treatment": "Improve field drainage and use a recommended copper-based bactericide as per label instructions.",
        "prevention": "Use clean seed, avoid excessive nitrogen, and reduce standing water when disease pressure is high.",
        "severity": "High",
        "affected_crops": ["Rice", "Maize"]
    },
    "Brown Spot": {
        "description": "Brown circular lesions with yellow halo are visible on the leaf surface.",
        "treatment": "Use a recommended fungicide such as Mancozeb if the spread is active and improve potassium nutrition.",
        "prevention": "Maintain balanced fertilization and avoid prolonged stress.",
        "severity": "Moderate",
        "affected_crops": ["Rice"]
    },
    "Leaf Blast": {
        "description": "Spindle or diamond-shaped lesions with grey center and brown edge are visible.",
        "treatment": "Use a suitable fungicide quickly and avoid excess nitrogen application.",
        "prevention": "Prefer resistant varieties and avoid dense crop canopy.",
        "severity": "High",
        "affected_crops": ["Rice", "Wheat"]
    },
    "Powdery Mildew": {
        "description": "White powder-like fungal growth is visible on leaf surfaces.",
        "treatment": "Use sulfur or another recommended fungicide and improve air circulation.",
        "prevention": "Avoid prolonged leaf wetness and maintain proper spacing.",
        "severity": "Moderate",
        "affected_crops": ["Wheat", "Pulses", "Vegetables"]
    },
    "Early Blight": {
        "description": "Dark concentric target-like lesions are visible on older leaves.",
        "treatment": "Remove badly affected leaves and use a suitable fungicide if spread continues.",
        "prevention": "Practice crop rotation and avoid overhead irrigation.",
        "severity": "Moderate",
        "affected_crops": ["Tomato", "Potato", "Brinjal"]
    },
    "Nitrogen Deficiency": {
        "description": "Lower leaves appear pale yellow, suggesting nutrient stress rather than infection.",
        "treatment": "Apply recommended nitrogen fertilizer dose based on crop stage and soil test status.",
        "prevention": "Follow a balanced nutrient schedule and test soil before sowing.",
        "severity": "Low",
        "affected_crops": ["Rice", "Maize", "Cotton", "All crops"]
    },
    "Stem Rust": {
        "description": "Reddish-brown rust-like pustules are visible, indicating possible fungal infection.",
        "treatment": "Use a recommended systemic fungicide promptly and monitor nearby plants.",
        "prevention": "Use resistant varieties and avoid delayed sowing where rust is common.",
        "severity": "High",
        "affected_crops": ["Wheat", "Barley"]
    },
}


class DiseaseDetectionService:
    def __init__(self):
        self.model = None
        self._load_model()

    def _load_model(self):
        model_path = os.environ.get("DISEASE_MODEL_PATH", "models/disease_model.pt")
        if os.path.exists(model_path):
            try:
                import torch
                self.model = torch.load(model_path, map_location="cpu")
                self.model.eval()
            except Exception:
                self.model = None

    def _analyze_image(self, image_path: str) -> tuple[str, float]:
        img = Image.open(image_path).convert("RGB").resize((224, 224))
        arr = np.array(img, dtype=np.float32)

        r, g, b = arr[:, :, 0], arr[:, :, 1], arr[:, :, 2]
        mean_r, mean_g, mean_b = float(r.mean()), float(g.mean()), float(b.mean())
        color_std = float(arr.std())
        green_ratio = float((g > (r + 10)).mean())
        yellow_ratio = float(((r > 125) & (g > 115) & (b < 110)).mean())
        white_ratio = float(((r > 180) & (g > 180) & (b > 180)).mean())
        dark_ratio = float(((r < 90) & (g < 90) & (b < 90)).mean())
        red_ratio = float(((r > g + 20) & (r > b + 20)).mean())

        if white_ratio > 0.18:
            disease = "Powdery Mildew"
            confidence = min(0.96, 0.72 + white_ratio)
        elif green_ratio > 0.52 and yellow_ratio < 0.10 and color_std < 52:
            disease = "Healthy"
            confidence = min(0.97, 0.78 + green_ratio * 0.25)
        elif yellow_ratio > 0.12 and mean_g < 145:
            disease = "Bacterial Leaf Blight"
            confidence = min(0.95, 0.70 + yellow_ratio + dark_ratio * 0.2)
        elif red_ratio > 0.22 and mean_r > 120:
            disease = "Stem Rust"
            confidence = min(0.94, 0.68 + red_ratio)
        elif dark_ratio > 0.16 and color_std > 62:
            disease = "Early Blight"
            confidence = min(0.93, 0.69 + dark_ratio)
        elif yellow_ratio > 0.08 and color_std > 55:
            disease = "Brown Spot"
            confidence = min(0.91, 0.67 + yellow_ratio + color_std / 500)
        elif mean_g < 105:
            disease = "Nitrogen Deficiency"
            confidence = min(0.89, 0.66 + (105 - mean_g) / 120)
        else:
            disease = "Leaf Blast"
            confidence = min(0.92, 0.68 + color_std / 300)

        return disease, round(confidence, 4)

    def _get_alternatives(self, primary: str) -> list:
        candidates = [name for name in DISEASE_KNOWLEDGE if name != primary][:2]
        probabilities = [0.12, 0.07]
        return [{"disease": name, "probability": prob} for name, prob in zip(candidates, probabilities)]

    async def analyze(self, image_path: str, language: str = "en") -> dict:
        disease_name, confidence = self._analyze_image(image_path)
        info = DISEASE_KNOWLEDGE.get(disease_name, DISEASE_KNOWLEDGE["Healthy"])

        result = {
            "disease_name": disease_name,
            "confidence": confidence,
            "confidence_percent": round(confidence * 100, 1),
            "severity": info["severity"],
            "description": info["description"],
            "treatment": info["treatment"],
            "prevention": info["prevention"],
            "affected_crops": info["affected_crops"],
            "is_healthy": disease_name == "Healthy",
            "alternative_diagnoses": self._get_alternatives(disease_name),
        }

        if language != "en":
            result["disease_name"] = await translate_text(result["disease_name"], language)
            result["description"] = await translate_text(result["description"], language)
            result["treatment"] = await translate_text(result["treatment"], language)
            result["prevention"] = await translate_text(result["prevention"], language)
            result["affected_crops"] = [await translate_text(item, language) for item in result["affected_crops"]]
            for item in result["alternative_diagnoses"]:
                item["disease"] = await translate_text(item["disease"], language)

        return result
