Downloads/agrimind-ai/agrimind/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
Pillow==10.3.0
numpy==1.26.4
scikit-learn==1.4.2
joblib==1.4.2
faiss-cpu==1.8.0
sentence-transformers==2.7.0
transformers==4.41.0
torch==2.3.0
torchvision==0.18.0
openai==1.30.0
google-generativeai==0.5.4
deep-translator==1.11.4
gTTS==2.5.1
SpeechRecognition==3.10.4
pydantic==2.7.1
python-dotenv==1.0.1
httpx==0.27.0
aiofiles==23.2.1
pandas==2.2.2
langchain==0.2.0
langchain-community==0.2.0
